Rename to ProgramMainContent and ProgramSidebar to avoid confusion
