+++ src/components/common/layout/MainContent.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 const MainContent = props => (
-  <div className="col-xs-12 col-lg-7">
+  <article className="col-xs-12 col-lg-7">
     {props.children}
-  </div>
+  </article>
 );
 
 MainContent.propTypes = {
