+++ src/components/common/layout/Sidebar.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 const Sidebar = props => (
-  <div className="col offset-lg-1">
+  <aside className="col offset-lg-1">
     {props.children}
-  </div>
+  </aside>
 );
 
 Sidebar.propTypes = {
